qrouting: Drop the gateway disabling concept
A gateway should _only_ be discounted from routing if it is dead (no
ping replies!), and not because it received a poor rating from qrouting.
On such a worst case, it is better to simply demote the gateway to the
last position in the sorted list.  This way, we don't risk breaking
connectivity.